requirements.txt
Flask==3.0.2
Pillow==10.2.0
pytesseract==0.3.10
qiskit==1.0.2
qiskit-aer==0.13.3
google-generativeai==0.3.2 